clients: fix compiler warning in _do_test_parse_passwd_file()
Since "_i" is int type and G_N_ELEMENTS() is size_t, the compiler
can warn

    error: comparison of unsigned expression in '< 0' is always false [-Werror=type-limits]

Avoid the warning by casting.

+++ clients/common/tests/test-clients-common.c
@@ -300,7 +300,7 @@ test_client_import_wireguard_missing(void)
                 g_assert_cmpint(-1, ==, *_p_error_line);                                     \
             g_assert(!_local);                                                               \
                                                                                              \
-            for (_i = 0; _i < G_N_ELEMENTS(_values); _i++) {                                 \
+            for (_i = 0; _i < (int) G_N_ELEMENTS(_values); _i++) {                           \
                 const NMUtilsNamedValue *_n = &_values[_i];                                  \
                 const char *             _v;                                                 \
                                                                                              \
